src/pyproject.toml
[tool.poetry]
name = "note management app backend"
version = "0.1.0"
description = ""
authors = ["Daniil Savelev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
drf-exceptions-hog = "^0.3.0"
drf-yasg = "^1.21.5"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^3.14.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
